Create submit utility folder
